More cleanup

+++ src/main/java/extrabiomes/renderers/CustomDoorRender.java
@@ -39,10 +39,7 @@ public class CustomDoorRender implements ISimpleBlockRenderingHandler {
                             : block.getMixedBrightnessForBlock(renderer.blockAccess, x, y - 1, z));
             tessellator.setColorOpaque_F(f, f, f);
             renderer.renderFaceYNeg(
-                    block,
-                    (double) x,
-                    (double) y,
-                    (double) z,
+                    block, x, y, z,
                     renderer.getBlockIcon(block, renderer.blockAccess, x, y, z, 0));
             flag = true;
         }
@@ -53,10 +50,7 @@ public class CustomDoorRender implements ISimpleBlockRenderingHandler {
                             : block.getMixedBrightnessForBlock(renderer.blockAccess, x, y + 1, z));
             tessellator.setColorOpaque_F(f1, f1, f1);
             renderer.renderFaceYPos(
-                    block,
-                    (double) x,
-                    (double) y,
-                    (double) z,
+                    block, x, y, z,
                     renderer.getBlockIcon(block, renderer.blockAccess, x, y, z, 1));
             flag = true;
         }
@@ -64,28 +58,28 @@ public class CustomDoorRender implements ISimpleBlockRenderingHandler {
                 renderer.renderMinZ > 0.0D ? i1 : block.getMixedBrightnessForBlock(renderer.blockAccess, x, y, z - 1));
         tessellator.setColorOpaque_F(f2, f2, f2);
         IIcon iicon = renderer.getBlockIcon(block, renderer.blockAccess, x, y, z, 2);
-        renderer.renderFaceZNeg(block, (double) x, (double) y, (double) z, iicon);
+        renderer.renderFaceZNeg(block, x, y, z, iicon);
         flag = true;
         renderer.flipTexture = false;
         tessellator.setBrightness(
                 renderer.renderMaxZ < 1.0D ? i1 : block.getMixedBrightnessForBlock(renderer.blockAccess, x, y, z + 1));
         tessellator.setColorOpaque_F(f2, f2, f2);
         iicon = renderer.getBlockIcon(block, renderer.blockAccess, x, y, z, 3);
-        renderer.renderFaceZPos(block, (double) x, (double) y, (double) z, iicon);
+        renderer.renderFaceZPos(block, x, y, z, iicon);
         flag = true;
         renderer.flipTexture = false;
         tessellator.setBrightness(
                 renderer.renderMinX > 0.0D ? i1 : block.getMixedBrightnessForBlock(renderer.blockAccess, x - 1, y, z));
         tessellator.setColorOpaque_F(f3, f3, f3);
         iicon = renderer.getBlockIcon(block, renderer.blockAccess, x, y, z, 4);
-        renderer.renderFaceXNeg(block, (double) x, (double) y, (double) z, iicon);
+        renderer.renderFaceXNeg(block, x, y, z, iicon);
         flag = true;
         renderer.flipTexture = false;
         tessellator.setBrightness(
                 renderer.renderMaxX < 1.0D ? i1 : block.getMixedBrightnessForBlock(renderer.blockAccess, x + 1, y, z));
         tessellator.setColorOpaque_F(f3, f3, f3);
         iicon = renderer.getBlockIcon(block, renderer.blockAccess, x, y, z, 5);
-        renderer.renderFaceXPos(block, (double) x, (double) y, (double) z, iicon);
+        renderer.renderFaceXPos(block, x, y, z, iicon);
         flag = true;
         renderer.flipTexture = false;
         return flag;
